restructure handling
